+++ pre_commit_hooks/ankidg_update_templates/__init__.py
@@ -1,0 +1,7 @@
+### GEN-INIT: TEMPLATE-START ###
+
+__all__ = ["ROOT", "TEMPLATEDIR"]
+
+from .dirs import ROOT, TEMPLATEDIR
+
+### GEN-INIT: TEMPLATE-CLOSE ###