+++ pre_commit_hooks/ankidg_update_templates/dirs.py
@@ -1,0 +1,6 @@
+from pathlib import Path
+
+__all__ = ["ROOT", "TEMPLATEDIR"]
+
+ROOT = Path(__file__).parent.parent.parent.resolve()
+TEMPLATEDIR = ROOT / "templates"