+++ src/anki_deck_gen/app/__init__.py
@@ -4,13 +4,18 @@ __all__ = [
     "AppVersion",
     "AsciiVersion",
     "CoreVersion",
+    "Level",
     "OdysseeVersion",
     "Version",
     "VersionAnnotation",
     "app",
+    "configure_logger",
+    "logger",
+    "parse_level",
 ]
 
 from .app import app
+from .logger import Level, configure_logger, logger, parse_level
 from .versions import (
     AppVersion,
     AsciiVersion,